+++ packages/editor/editor-core/src/plugins/table/utils/fix-tables.ts
@@ -1,6 +1,8 @@
 import { Transaction } from 'prosemirror-state';
 import { Node as PMNode } from 'prosemirror-model';
 
+import { parseDOMColumnWidths } from '../utils';
+
 const fixTable = (
   table: PMNode,
   tablePos: number,
@@ -69,4 +71,43 @@ export const fixTables = (tr: Transaction): Transaction => {
     }
   });
   return tr;
+};
+
+// When we get a table with an 'auto' attribute, we want to:
+// 1. render with table-layout: auto
+// 2. capture the column widths
+// 3. set the column widths as attributes, and remove the 'auto' attribute,
+//    so the table renders the same, but is now fixed-width
+//
+// This can be used to migrate table appearances from other sources that are
+// usually rendered with 'auto'.
+//
+// We use this when migrating TinyMCE tables for Confluence, for example:
+// https://pug.jira-dev.com/wiki/spaces/AEC/pages/3362882215/How+do+we+map+TinyMCE+tables+to+Fabric+tables
+export const fixAutoSizedTable = (
+  tr: Transaction,
+  node: PMNode,
+  table: HTMLTableElement,
+  basePos: number,
+) => {
+  const colWidths = parseDOMColumnWidths(table);
+
+  node.forEach((rowNode, rowOffset, i) => {
+    rowNode.forEach((colNode, colOffset, j) => {
+      const pos = rowOffset + colOffset + basePos + 2;
+
+      tr.setNodeMarkup(pos, undefined, {
+        ...colNode.attrs,
+        colwidth: colWidths.width(j, colNode.attrs.colspan).map(Math.round),
+      });
+    });
+  });
+
+  // clear autosizing on the table node
+  return tr
+    .setNodeMarkup(basePos, undefined, {
+      ...node.attrs,
+      __autoSize: false,
+    })
+    .setMeta('addToHistory', false);
 };